refactor(hooks): move hooks folder to .hooks
We have choosed to do in that way to improve the code
organization
